factored out gb.cpp
