+++ weld_rt/cpp/Makefile
@@ -28,13 +28,16 @@ vb.o: vb.cpp
 dict.o: dict.cpp
 	${CLANG} ${CFLAGS} -c $< -o $@
 
+gb.o: gb.cpp
+	${CLANG} ${CFLAGS} -c $< -o $@
+
 merger.o: merger.cpp
 	${CLANG} ${CFLAGS} -c $< -o $@
 
 runtime.o: runtime.cpp
 	${CLANG} ${CFLAGS} -c $< -o $@
 
-weldrt.o: vb.o dict.o merger.o runtime.o inline.o
+weldrt.o: vb.o dict.o gb.o merger.o runtime.o inline.o
 	ld -r $^ -o $@
 
 ${LIB}: weldrt.o
